Red Autonomous
Red Park Files Here are made for the autonomous portion of competition.
Each file name coordinates with the starting position of the Robot on the field. For example, "RedParkFrontLeft" means that the robot starts in the front of the field (closest to paper drone landing area), stops at the left side tape mark on the field and then travels to the backstage area. 
Files "BackstageFront" or "BackstageBack" still coordinate with the front or back location on the field, but they do not stop at a specific tape mark and go directly to the backstage area. 

All of these codes end in the backstage area.